Extract shared utilities and clean up code duplication
- Extract relativeTime, getCategoryIcon, isInSuggestionPhase, BADGE_COLORS,
  getResultBadge to lib/pollListUtils.ts (was duplicated in 3 files)
- Extract loadVotedPolls to lib/votedPollsStorage.ts (was duplicated in 3 files)
- Add showForkButton prop to FollowUpModal instead of duplicating the component
- Remove local duplicates from thread page and ThreadList
- Remove unused imports (getUserName, FollowUpModal duplicate, useMemo, useCallback)
- Remove unused allPolls state from thread page
- Pre-compute sort keys in threadUtils (avoid Date parsing in sort comparator)
- Optimize backend voter_names: extract from already-fetched votes first,
  only query DB for remaining open polls
